tetragon: Add support for path offload
Adding support to retrieve full path (bigger path for !large programs)
by moving the retrieval code into separate programs where we have space
for all the instructions to parse the path.

The idea is to tail call in path retrieval code, which then tail calls
recursively or finish.

The maximum allowed path size is 4096, so we allow 512 path entry iterations
and 8 recursive tail calls.

We have to limit that for !large programs (v4.19) to 32 path entry iterations,
so there's 256 limit on path components for !large programs, which is
still improvement to current 11.

We offload the path retrieval on kernels before v6.1 kernels, because
they already have big loops support.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>