ci: Replace generated-files workflow with generated-docs
Now that code generation is checked in dedicated workflows, the only checks
remaining in generated-files are flags and metrics references. Let's rename
the workflow and the job to better represent its purpose. Like before, there is
a single job, but since two generated docs are unrelated, let's not fail after
the first failed step.

I adjusted the triggers, so that we run the workflow on relevant changes. In
particular, run it on changes to the generated files - before the entire docs
directory was skipped, so if someone made changes to one of the references
manually, the workflow wouldn't have caught it.

I also removed lint-metrics-md Makefile target and instead added an inline
check in the workflow, consistent with other similar checks.

Signed-off-by: Anna Kapuscinska <anna@isovalent.com>