tetragon/windows: Add a global exec collection object
This commit adds a global collection object to support process create and process exit events on Windows.
The ringbuffer maps on Windows currently have some issues with pinning. As a result we do not pin maps after program load.
In order to read from those maps we need to keep a global collection object alive and use the FD in the collection object to read from those maps.
This is a transitory change till pinning issues is fixed on Windows.
After which, this global object will be removed and we will use pinned maps to read ringbuf maps.

Signed-off-by: Anadi Anadi<aanadi@cisco.com>